finished player class for now. added vector2d class (no functionality required yet)

+++ src/math/Rectangle.hpp
@@ -2,7 +2,7 @@
 #include "SDL.h"
 
 namespace RectUtil {
-	void Translate(SDL_Rect& Rectangle, const int& t_x, const int& t_y);
+	void Translate(SDL_Rect& Rectangle, const float& t_x, const float& t_y);
 	void Scale(SDL_Rect& Rectangle, const float& s_x, const float& s_y);
 	bool CheckCollision(const SDL_Rect& R1, const SDL_Rect& R2);
 }
